# Optimized `stolarsky_mean`